Refactored to multiple projects
BattleBukkitLib v1.0.6

There major problem was that I could not import BattleBukkitLib into the
Netbeans IDE. I think it's important to support multiple IDEs. This new
POM setup should also be supported by Eclipse. Also, another reason why this is
important... is because it's important to be able to visualize and see all
the compiler errors. And IDEs make it a little easier to visualize these
errors over the command-line.

The obvious downside is that now there's multiple projects to work with
rather than just one project... But this mirrors the build process
anyways. Also, hopefully it'll make it easier to update in the future:
because alls you have to do is copy/paste the previous vX_Y_Z module, then edit
the pom.xml to the new version, then edit the imports and check for
compiler errors.

Like I said: Hopefully, this makes the dependency-graph and the build
process more clear (and simplier, easier to read POMs). If not, lemme
know.

There are now 3 conceptual layers:

1. The interface layer (modules/api)
2. The implementation layer (modules/vX_Y_Z)
3. The runtime Utility layer (modules/BattleBukkitLib)

The interface layer is the abstraction layer that defines all the
interfaces (what we want to do).

The implementation layer consists of multiple implementations that depend
on the interface layer (modules/api) and implements the interfaces.

The runtime Utility layer is a set of static utility classes. Each static
method is powered by an interface. And any calls to these Utility methods
are merely forwarded to the interface that powers it.

The runtime Utility layer also consist of a set of internal factories that
choose (at runtime) which implementation will be used.